feat(container): add a container
We add a container with bottlejs, to make
adding dependencies easier in the future.

I wanted to add bdd test, that is one of
the reasons that the container is added,
so I could esaliy change between the real
services and a memory implementation to use
in the tests. Sadly at the end I didn't
had time to finish everything and I had to
skip bdd tests.