feat(dataBase): introduce external service
After spending several hours reading The Movie DB
documentation, trying to figure out how to get more than
20 lines of data at once or how to get better-filtered data,
I had to conclude that there isn't a better way than using
the top_rated API.
For a brief second, I saw some hope that I can use the
discover API, but I would need to know the exact vote
average of the 210th top-rated movie, but the vote
average that we get is a rounded number, so the 210th
together with 30 or more movies, shows the same
number. Although, it is rounded probably after the
database request because if I changed the decimals, it
changed the result. Sadly even though it could work, it
would be a guessing game, and the average could change
any day, so it is not a maintainable solution. That is why I
had no other choice than to use the top_rated API.
It is quite bad news because it means we will have to send
several requests at once and gather all 210 movies to be
able to filter it.