feat(MovieState): add internal service
After looking through The Movie DB documentation, we
saw that the length of the movie is not included in the
data when we request the top-rated movies. We have to
gather this information one by one with the details API to
show every data requested on the top-rated list page.
Maybe we could make fewer calls at once by not
gathering the details only for the movies that are in the
filter, but that way, we can send a lot more requests. This
way, we can get all the data at once, and we don't have to
request them again until the page gets refreshed.
I did not use fire and forget methods, even though it might
make things faster on the frontend side, because I am not
familiar with The Movie DB scalability, and I don't want to
flood their server.

The functions are containing some general logic, that should
be refactored and separated from business logic, but I run
out of time to do so.