feat(pages/main): add main page
This page exists to have a baseline. To show error
messages or loading screens. Also, since we have to
send so many requests at once we don't want to resend
them every time the user visits a different page on our
site. This way, we load the data once or if the user
refreshes the site but not at page changes.