feat(lib/messages): add function
This function should decide if the response of
an http request is successful or not. If it get's
an error, it will forward it into the messageState,
so we can show it to the user. The handler and the
http service error response was built in mind of
The Movie DB, and their status message.